Fix XML comments not seen on base generic class (.Net 4.0+ & Standard 2.0)
Partially fixes the issue #8 where XML comments on members inherited from a base class that uses generics are not seen/returned.

The fix also improves some side cases with special method parameters (ref parameters, array parameters, pointers, etc..).

The fix is not 100% but a good step forward. It is OK for .Net Framework 4.0+ and .Net Standard 2.0.

The fix is not complete as the base generic class problem cannot be fixed for .Net Standard 1.0 (the Reflection API lacks the required methods and metadata in .Net Standard 1.0).

As for the Mono.Cecil XmlDocument extraction, it still uses the old code. But a port from the New Reflection Code should be doable without too much complications.

A new unit test verifies that the problem is fixed.